Reorganizar a parte de gerar a map de variáveis
